chore: Sensor related packages refactory
