+++ sensors/artifacts/configmap_test.go
@@ -1,4 +1,4 @@
-package store
+package artifacts
 
 import (
 	"context"
