+++ sensors/artifacts/resource_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package store
+package artifacts
 
 import (
 	"testing"
