+++ sensors/artifacts/url_test.go
@@ -1,4 +1,4 @@
-package store
+package artifacts
 
 import (
 	"fmt"
